[Fix] Frame

+++ KikukeLib/lsys/sptime.c
@@ -5,7 +5,7 @@ double getTimeDist(const struct timeval* start, const struct timeval* end)
     double dist;
 
     dist = ((*end).tv_sec - (*start).tv_sec);
-    dist += ((*end).tv_usec - (*start).tv_usec) / 1000000;
+    dist += (double)((*end).tv_usec - (*start).tv_usec) / 1000000;
 
     return dist;
 }
